// src/types/Global.ts
export type Users = {
  name: string;
  login: string;
  repos_url: string;
  // id: 130059187;
  // node_id: 'U_kgDOB8CLsw';
  // avatar_url: 'https://avatars.githubusercontent.com/u/130059187?v=4';
  // gravatar_id: '';
  // url: 'https://api.github.com/users/valmirpst';
  // html_url: 'https://github.com/valmirpst';
  // followers_url: 'https://api.github.com/users/valmirpst/followers';
  // following_url: 'https://api.github.com/users/valmirpst/following{/other_user}';
  // gists_url: 'https://api.github.com/users/valmirpst/gists{/gist_id}';
  // starred_url: 'https://api.github.com/users/valmirpst/starred{/owner}{/repo}';
  // subscriptions_url: 'https://api.github.com/users/valmirpst/subscriptions';
  // organizations_url: 'https://api.github.com/users/valmirpst/orgs';
  // events_url: 'https://api.github.com/users/valmirpst/events{/privacy}';
  // received_events_url: 'https://api.github.com/users/valmirpst/received_events';
  // type: 'User';
  // site_admin: false;
  // company: 'UTFPR';
  // blog: 'https://valmirpst.github.io/portfolio/';
  // location: 'Campo Mourão - PR';
  // email: null;
  // hireable: null;
  // bio: 'Técnico Integrado em Informática (3/4)';
  // twitter_username: null;
  // public_repos: 8;
  // public_gists: 0;
  // followers: 5;
  // following: 7;
  // created_at: '2023-04-06T11:36:10Z';
  // updated_at: '2024-06-22T12:48:42Z';
};

export type Repos = {
  name: string;
  description: string;
  language: string;
};